Fix packaging issue
Compilation was failing when archiving. This problem appeared because of the model asking for code generation. This code generation was in fact not required, it merely existed to suppress an Xcode build system warning. It is therefore better to have the code compile and archive correctly, and to live with the warning (or to switch back to the old build system).