Use single request when sending history updates
No 50 limit exists anymore. Sending several requests in parallel might confuse the server and lead to dropped entries. The safest is to send everything in a batch (the server should return an error if the insertion failed, and sync would occur probably with the next attempt).

This fixes the large history test which was sometimes (well most of the time) failing for this reason.